Created sinlge_dimensional_PCA.ipynb
this is file that is making 1D PCA plots.

(1) it creates PCA like arrays of centered spike times for all neurons by loading in the df_units.json file
(2) it then plots that on a 1D plot. where we can see how all neurons are firing at any single time bin during the recording

This is interesting because a 3D PCA is just this done but with all the time bins at once.